Add force_active Optional arg to playback_resumed and add logout method
The web client uses the 'forceActive' param to force the current session to be the active session in case the user is already listening on another device. This handles STREAM_VIOLATION errors for paid users. Logging out helps prevent INVALID_TOKEN errors for paid users. In the future when/if we store the auth token it may not be needed.